+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 =====
 
 * Aandlers for password restriction configurations.
+* Spanish (es_ES/es_419) translations for the xblock.
 
 0.1.0 – 2024-03-26
 **********************************************